rework: big reorganization of how the job and jobManager are used
- reimagine repartition between parent and derives classes
- move jobManagers into their own folder
- refactor some code
- remove unused methods
