+++ EasyCLI/Commands/EditCommand.cs
@@ -1,8 +1,8 @@
 using EasyCLI.Commands.CommandFeatures;
 using EasyCLI.Commands.CommandFeatures.CommandArgType;
 using EasyCLI.Localization;
-using EasyLib;
 using EasyLib.Enums;
+using EasyLib.JobManager;
 
 namespace EasyCLI.Commands;
