+++ EasySave.Tests/EasyLib/JobManagerTests.cs
@@ -1,5 +1,5 @@
-using EasyLib;
 using EasyLib.Enums;
+using EasyLib.JobManager;
 
 namespace EasySave.Tests.EasyLib;
